+++ project/src/components/login/login.tsx
@@ -1,8 +1,7 @@
-import Sprite from '../main/sprite/sprite';
-import Logo from '../main/logo/logo';
+import Sprite from '../main/sprite';
+import Logo from '../main/logo';
 
 function LogIn(): JSX.Element {
-
   return (
     <>
