+++ project/src/components/login/user-menu.tsx
@@ -1,5 +1,4 @@
 function UserMenu(): JSX.Element {
-
   return (
     <nav className="header__nav">
       <ul className="header__nav-list">
