+++ project/src/components/main/cards.tsx
@@ -1,6 +1,6 @@
-import Card from '../card/card';
-import { Offer } from '../../../../src/types/types';
-import { Modificator } from '../../const';
+import Card from './card';
+import { Offer } from '../../types/types';
+import { Modificator } from '../const';
 
 type cardsProps = {
   offers: Offer[],
@@ -11,7 +11,6 @@ type OnHover = {
 }
 
 function Cards( { offers, onHover }: cardsProps & OnHover): JSX.Element {
-
   return (
     <div className="cities__places-list places__list tabs__content">
       { offers.map(({ id, ...rest }) => <Card id={ id } key={ id } onHover={ onHover } { ...rest } modificator={ Modificator.MAIN } />) }
