+++ project/src/components/main/footer.tsx
@@ -1,7 +1,6 @@
 import { Link } from 'react-router-dom';
 
 function Footer(): JSX.Element {
-
   return (
     <footer className="footer container">
       <Link className="footer__logo-link" to='/'>
