+++ project/src/components/main/header.tsx
@@ -1,8 +1,7 @@
-import UserMenu from '../../login/user-menu/user-menu';
-import Logo from '../logo/logo';
+import UserMenu from '../login/user-menu';
+import Logo from './logo';
 
 function Header(): JSX.Element {
-
   return (
     <header className="header">
       <div className="container">
