+++ project/src/components/main/logo.tsx
@@ -1,7 +1,6 @@
 import { Link } from 'react-router-dom';
 
 function Logo(): JSX.Element {
-
   return (
     <Link className="header__logo-link" to="/">
       <img
