+++ project/src/components/main/sort.tsx
@@ -1,5 +1,4 @@
 function Sort(): JSX.Element {
-
   return (
     <form className="places__sorting" action="#" method="get">
       <span className="places__sorting-caption">Sort by</span>
