+++ project/src/components/main/sprite.tsx
