+++ project/src/components/map/map.tsx
@@ -12,7 +12,6 @@ type MapProps = {
 }
 
 function Map({ offers, activeOffer }: MapProps): JSX.Element {
-
   const { city } = offers[0];
   const mapRef = useRef(null);
   const map = useMap(mapRef, city);
