+++ project/src/components/properties/price.tsx
@@ -3,7 +3,6 @@ type PriceProps = {
 }
 
 function Price({ price }: PriceProps): JSX.Element {
-
   return (
     <div className="property__price">
       <b className="property__price-value">&euro;{ price }</b>
