+++ project/src/components/properties/reviews.tsx
@@ -6,7 +6,6 @@ type ReviewsProps = {
 }
 
 function Reviews({ reviews }: ReviewsProps): JSX.Element {
-
   return (
     <>
       <h2 className="reviews__title">Reviews &middot; <span className="reviews__amount">{reviews.length}</span></h2>
