+++ project/src/utils/utils.tsx
@@ -25,10 +25,11 @@ export const transformRating = (rating: number): string => `${ rating / RatingTo
 
 const ICON_WIDTH = 40;
 const ICON_HEIGHT = 13;
+const ICON_HEIGHT2 = 26;
 
 export const iconChanger = (change: string): Icon<IconOptions> | DivIcon => leaflet.icon({
   iconUrl: change,
-  iconSize: [ICON_HEIGHT * 2, ICON_WIDTH],
+  iconSize: [ICON_HEIGHT2, ICON_WIDTH],
   iconAnchor: [ICON_HEIGHT, ICON_WIDTH],
 });
 
@@ -36,9 +37,11 @@ export const articleClass = (param: string): string=> {
   if (param === Modificator.PROPERTIES) {
     return 'near-places__card place-card';
   }
+
   if (param === Modificator.FAVORITES) {
     return 'favorites__card place-card';
   }
+
   return 'cities__place-card place-card';
 };
 
@@ -46,8 +49,10 @@ export const imageWrapper = (param: string): string => {
   if (param === Modificator.PROPERTIES) {
     return 'near-places__image-wrapper place-card__image-wrapper';
   }
+
   if (param === Modificator.FAVORITES) {
     return 'favorites__image-wrapper place-card__image-wrapper';
   }
+
   return 'cities__image-wrapper place-card__image-wrapper';
 };
